- fragment_register.xml: New UI elements added. Including red error icons and red error labels for the password and email EditText views. These UI elements are hidden by default and will appear when an invalid password or email is entered. password must be at least 8 characters email must be a valid email address
- RegisterFragment.kt: New addTextChangedListener lambdas are attached to the passwordEditText and the emailEditText. These listeners control the visibility of the new error UI elements.
Coded logic to handle the submit button. We will cache the values from the username, password, and email EditText's and attempt to sign up a new user with AWS Cognito. If successful we immediately sign the user in with the same cached credentials. If sign in is successful UserData.loggedIn is set to true, a toast is displayed, and we navigate back to FirstFragment.
If an error occurs either when signing up, or signing in a toast is displayed with the error message.

- strings.xml/colors.xml/error.xml: Added new strings for the toast messages and error labels. Added a red color for the register fragments new error UI elements. Error.xml is the error icon being used.